Remove requirements.txt
Earlier I used a pip-tools requirements.in -> requirements.txt workflow. But
there were some supplementary tools which are required for development viz.,
zstd and libarchive for which pure python workflow was not enough.

Essential dependencies are now specified in setup.cfg and are consumed by the test
environment (CI: github-actions). For the runtime environment we use conda and
pin dependencies hard using conda's environment.yml file